fixed a lot of issues

+++ src/Headers/Board.h
@@ -56,7 +56,7 @@ public:
     void print();
     bool isPositionInBounds(pair<int,int> position);
     shared_ptr<Piece> addPiece(Piece::Type type, Piece::Color color, pair<int,int> position);
-    float evaluate();
+    double evaluate();
     int invertRow(int row);
     string toString();
